Added buildings awarding students money and converted buildings students to f64 from i32.

+++ src/main.rs
@@ -20,7 +20,7 @@ use macroquad::ui::{
 };
 struct Building {
     name: &'static str, // The type of building
-    students: i32,    // Students per Second that this building generates
+    students: f64,    // Students per Second that this building generates
     perk_points: i32,// Number of perk points awarded by purchasing this building
     price: i64, // Price to purchase building
     description: &'static str,
